cleanup removeExtension and addMode
