+++ platform/cli/src/commands/removeExtension.js
@@ -3,72 +3,13 @@ import Listr from 'listr';
 
 import {
   uninstallNPMPackage,
-  readPluginConfigFile,
-  removeExtensionFromConfig,
-  writePluginConfigFile,
+  throwIfExtensionUsedByInstalledMode,
+  removeExtensionFromConfigFile,
   validateExtensionYarnInfo,
-  getYarnInfo,
 } from './utils/index.js';
 
 export default async function removeExtension(packageName) {
   console.log(chalk.green.bold(`Removing ohif-extension ${packageName}...`));
-
-  async function removeExtensionFromConfigFile() {
-    const pluginConfig = readPluginConfigFile();
-
-    // Note: if file is not found, nothing to remove.
-    if (pluginConfig) {
-      removeExtensionFromConfig(pluginConfig, { packageName });
-      writePluginConfigFile(pluginConfig);
-    }
-  }
-
-  async function throwIfPackageUsedByAMode(packageName) {
-    const pluginConfig = readPluginConfigFile();
-
-    if (!pluginConfig) {
-      // No other modes, not in use
-      return false;
-    }
-
-    const { modes } = pluginConfig;
-
-    const modesUsingExtension = [];
-
-    debugger;
-
-    for (let i = 0; i < modes.length; i++) {
-      const mode = modes[i];
-      const modePackageName = mode.packageName;
-      const yarnInfo = await getYarnInfo(modePackageName);
-
-      const peerDependencies = yarnInfo.peerDependencies;
-
-      if (!peerDependencies) {
-        continue;
-      }
-
-      if (Object.keys(peerDependencies).includes(packageName)) {
-        modesUsingExtension.push(modePackageName);
-      }
-    }
-
-    if (modesUsingExtension.length > 0) {
-      let modesString = '';
-
-      modesUsingExtension.forEach(packageName => {
-        modesString += ` ${packageName}`;
-      });
-
-      const error = new Error(
-        `${chalk.yellow.red(
-          'Error'
-        )} ohif-extension ${packageName} used by installed modes:${modesString}`
-      );
-
-      throw error;
-    }
-  }
 
   const tasks = new Listr(
     [
@@ -78,7 +19,8 @@ export default async function removeExtension(packageName) {
       },
       {
         title: `Checking if ${packageName} is in use by an installed mode`,
-        task: async () => await throwIfPackageUsedByAMode(packageName),
+        task: async () =>
+          await throwIfExtensionUsedByInstalledMode(packageName),
       },
       {
         title: `Uninstalling npm package: ${packageName}`,
@@ -86,15 +28,13 @@ export default async function removeExtension(packageName) {
       },
       {
         title: 'Removing ohif-extension from the configuration file',
-        task: removeExtensionFromConfigFile,
+        task: async () => removeExtensionFromConfigFile(packageName),
       },
     ],
     {
       exitOnError: true,
     }
   );
-
-  // TODO -> Warn of extension removal when an extension is removed that a mode is using?
 
   await tasks
     .run()
