+++ platform/viewer/pluginConfig.json
@@ -18,20 +18,12 @@
         },
         {
             "version": "4.12.11"
-        },
-        {
-            "packageName": "@ohif-test/extension-clock",
-            "version": "1.0.2"
         }
     ],
     "modes": [
         {
             "packageName": "@ohif/mode-longitudinal",
             "version": "0.0.1"
-        },
-        {
-            "packageName": "@ohif-test/mode-clock",
-            "version": "1.0.5"
         }
     ]
 }
